refactor: classroom 도메인을 기능별 패키지 구조로 분리
