+++ app/domains/classroom/controller/teacher_class_router.py
@@ -12,11 +12,11 @@ from fastapi import APIRouter, Depends, HTTPException, Query, status
 
 from app.domains.auth.dependencies import get_current_teacher
 from app.domains.auth.models import User
-from app.domains.classroom.dependencies import get_classroom_service
-from app.domains.classroom.service import ClassroomService
+from app.domains.classroom.dependency.dependencies import get_classroom_service
+from app.domains.classroom.service.service import ClassroomService
 from app.domains.progress.dependencies import get_learning_record_service
 from app.domains.progress.service import LearningRecordService
-from app.domains.classroom.schemas import (
+from app.domains.classroom.schema.schemas import (
     AddStudentRequest,
     CreateClassRequest,
     TeacherClassesResponse,
