+++ app/domains/progress/teacher_router.py
@@ -11,8 +11,8 @@ from fastapi import APIRouter, Depends, HTTPException, Query, status
 
 from app.domains.auth.dependencies import get_current_teacher
 from app.domains.auth.models import User
-from app.domains.classroom.dependencies import get_classroom_service
-from app.domains.classroom.service import ClassroomService
+from app.domains.classroom.dependency.dependencies import get_classroom_service
+from app.domains.classroom.service.service import ClassroomService
 from app.domains.progress.dependencies import get_learning_record_service
 from app.domains.progress.schemas import (
     LearningRecordResponse,
