+++ tests/classroom/test_classroom_service.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
 
 from app.domains.auth.models import User
-from app.domains.classroom.service import ClassroomService
+from app.domains.classroom.service.service import ClassroomService
 
 
 class FakeClassroomRepository:
