+++ tests/structure/test_domain_structure.py
@@ -14,13 +14,14 @@ def test_domain_packages_are_importable():
         "app.domains.auth.router",
         "app.domains.auth.schemas",
         "app.domains.auth.service",
-        "app.domains.classroom.schemas",
-        "app.domains.classroom.teacher_class_router",
+        "app.domains.classroom.controller.teacher_class_router",
+        "app.domains.classroom.controller.student_class_router",
         "app.domains.progress.teacher_router",
-        "app.domains.classroom.dependencies",
+        "app.domains.classroom.dependency.dependencies",
         "app.domains.classroom.models",
-        "app.domains.classroom.repository",
-        "app.domains.classroom.service",
+        "app.domains.classroom.repository.repository",
+        "app.domains.classroom.service.service",
+        "app.domains.classroom.schema.schemas",
         "app.domains.agent.schemas",
         "app.domains.agent.service",
         "app.domains.instruction.student_router",
